requirements.txt
streamlit==1.41.1
streamlit-authenticator==0.3.3
pymongo==4.10.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
